Port all code for loading geojson into typescript.
